Rewrote follower spawning/saving
Originally was using a scriptable object for follower management, but since it was a single source it didn't work for multiple saves. Created a manager script that will generate new followers and saves their state with itself to restore during loading.

Need to look into breaking up SaveableEntity into another part so that data's not getting doubled.